+++ src/utils.js/render.js
@@ -9,7 +9,7 @@ export const createElement = (template) => {
   const newElement = document.createElement('div');
   newElement.innerHTML = template;
 
-  return newElement.firstChild;
+  return newElement.firstElementChild;
 };
 
 export const render = (container, child, place) => {
